+++ src/test/types/date.spec.ts
@@ -10,14 +10,18 @@ describe("DateType", function () {
       name: "new Date(0)",
       value: new Date(0),
       valid: true,
-      serialized: {
-        json: {
-          canonical: "1970-01-01T00:00:00.000Z",
-          values: [
-            {value: 0, valid: true},
-            {value: null, valid: false}
-          ]
-        }
+      output: {
+        json: "1970-01-01T00:00:00.000Z"
+      },
+      inputs: {
+        json: [
+          0
+        ]
+      },
+      invalidInputs: {
+        json: [
+          null
+        ]
       }
     },
     {name: 'new Date("1247-05-18T19:40:08.418Z")', value: new Date("1247-05-18T19:40:08.418Z"), valid: true},