+++ src/test/types/float64.spec.ts
@@ -9,13 +9,8 @@ describe("Float64Type", function () {
       name: "0",
       value: 0,
       valid: true,
-      serialized: {
-        "json-doc": {
-          canonical: 0,
-          values: [
-            {value: 0, valid: true}
-          ]
-        }
+      output: {
+        json: 0
       }
     },
     {name: "1", value: 1, valid: true},