+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Next
 
+- **[Feature]** Add the `rename` option to `DocumentType`. This allows to rename
+  the keys of the document similarly to the values of `SimpleEnumType`.
 - **[Internal]** Fix documentation generation with typedoc
+- **[Internal]** Improve support for the tests checking the output of `.write`
 
 # 0.5.0-alpha.3
 
-- **[Minor]** Add support for the `"qs"` format for [the `qs` package][npm-qs] (used by [`express`][npm-express])
+- **[Feature]** Add support for the `"qs"` format for [the `qs` package][npm-qs] (used by [`express`][npm-express])
 - **[Internal]** Create CHANGELOG.md
 
 [npm-express]:https://www.npmjs.com/package/expess